+++ app/src/main/java/nextstep/shoppingcart/ui/shoppinglist/ShoppingListColumn.kt
@@ -8,7 +8,7 @@ import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
-import nextstep.shoppingcart.ProductRepository
+import nextstep.shoppingcart.data.ProductRepository
 import nextstep.shoppingcart.model.Product
 import nextstep.shoppingcart.ui.theme.ShoppingCartTheme
