+++ app/src/main/java/nextstep/shoppingcart/ui/shoppinglist/ShoppingListScreen.kt
@@ -14,7 +14,7 @@ import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.tooling.preview.Preview
-import nextstep.shoppingcart.ProductRepository
+import nextstep.shoppingcart.data.ProductRepository
 import nextstep.shoppingcart.R
 import nextstep.shoppingcart.ui.theme.EerieBlack
 import nextstep.shoppingcart.ui.theme.ShoppingCartTheme
